+++ plugins/unplugin/src/bundlers/webpack/create.ts
@@ -45,6 +45,7 @@ const createPlugin = (createOptions?: CreatePluginOptions): Plugin => {
                 cwd,
                 include: options?.input?.include,
                 exclude: options?.input?.exclude,
+                tsconfig: options?.input?.tsconfig,
             });
 
         const logger: Logger = createLogger({