+++ tests/plugins/vite/tests/index.test.ts
@@ -27,7 +27,7 @@ beforeAll(async (): Promise<void> => {
 describe("vite test", (): void => {
     it("should build via Vite", async (): Promise<void> => {
         await build({
-            root: process.cwd(),
+            root: CWD,
             build: {
                 outDir: "dist",
                 minify: false,